requirements.txt
langchain==0.0.344
python-dotenv==1.0.0
python-telegram-bot==20.7
redis==5.0.1
gigachat